pyproject.toml
[project]
name = "graphiti-qdrant"
version = "0.1.0"
description = "Graphiti Qdrant Loader"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.4.1",
    "langchain-community>=0.4.1",
    "ipywidgets>=8.1.8",
    "langchain>=1.1.0",
    "langchain-qdrant>=1.1.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
    "fastmcp>=0.7.0",
]
